Add optimize mode (layout/yaw/both) and fix float64 promotion in the physics layer
GAConfig.optimize selects the decision variables: "both" (default, unchanged
behaviour), "layout" (yaw pinned to 0) or "yaw" (one fixed layout shared by the
whole population, seeded from seed_layout when given). In yaw mode the
projection chain runs once in init_population and project() short-circuits;
mutate() skips the kernels for whichever variables are frozen. Exposed as
--optimize on the CLI and as OPTIMIZE in example_optimizer.py.

Layout mode makes yaw identically 0, which lets FarmEvaluator.evaluate() skip
the Gauss deflection model entirely: theta_c0 is proportional to yaw, so
delta0 and mid_term vanish and only the ad + bd*dx offset survives. Gated on
ad == bd == 0 and on an explicit zero_yaw flag -- inferring it via
cp.all(yaw == 0) would force a device sync inside the fitness loop.

Chasing why that fast path was not bit-exact turned up cp.interp: it mirrors
np.interp and always returns float64, so Ct and power promoted axial induction,
TI, deflection, deficit and every (B, T, T) tensor downstream to float64. Same
trap from a np.exp scalar in the deflection model's E0 term. Casting both back
to float32 leaves AEP unchanged to float32 precision (the tables and layout
inputs are float32, so the extra bits carried no information) and all 12 FLORIS
comparison tests still pass.

Measured per generation at P=256, T=20, 12x11 wind rose:
    both    464 ms -> 185 ms
    layout  464 ms -> 110 ms
    yaw     464 ms -> 181 ms

test_ct_clipped asserted 0.0001 <= ct against a float64 literal; float32(0.0001)
is 9.9999997e-05, so the bounds are now taken in float32.

+++ packages/gpuwfarm-core/src/gpuwfarm_core/physics/turbine/power_curve.py
@@ -113,8 +113,12 @@ class TabulatedPowerCurve:
             Ct, clipped to [0.0001, 0.9999] — same as FLORIS
         """
         u_corr = self._density_corrected_velocity(u_eff)
-        ct = cp.interp(u_corr, self._ws_gpu, self._ct_gpu)
-        return cp.clip(ct, 0.0001, 0.9999)
+        # .astype(float32): cp.interp mirrors np.interp and always returns float64.
+        # Without the cast, Ct -- and through it axial induction, TI, deflection,
+        # deficit and every (B, T, T) tensor downstream -- silently runs in float64,
+        # doubling memory traffic for no accuracy gain (the tables are float32).
+        ct = cp.interp(u_corr, self._ws_gpu, self._ct_gpu).astype(cp.float32)
+        return cp.clip(ct, cp.float32(0.0001), cp.float32(0.9999))
 
     # ------------------------------------------------------------------
     # Axial induction (SimpleTurbine.axial_induction)
@@ -143,6 +147,6 @@ class TabulatedPowerCurve:
         u_corr = self._density_corrected_velocity(u_eff)
         # FLORIS uses fill_value=0.0 for out-of-bounds (SimpleTurbine.power)
         p_base = cp.interp(u_corr, self._ws_gpu, self._pow_gpu,
-                           left=cp.float32(0.0), right=cp.float32(0.0))
+                           left=cp.float32(0.0), right=cp.float32(0.0)).astype(cp.float32)
         yaw_loss = cp.cos(yaw) ** self.td.cosine_loss_exponent_yaw
         return p_base * yaw_loss
